test: SSEClient 테스트 helper 분리 + 응답 검증 테스트 분리
SwiftLint type_body_length error (>400 lines) 해소.

- 신규: SSEClientTestSupport.swift — TestStream / ProviderBuilder /
  StateRecorder / makeSSEClient / waitForCondition 공유 helper
- 신규: SSEClientResponseValidationTests.swift — 200/Content-Type 가드
  관련 케이스 (204 reject, wrong content-type reject, charset 허용)
- SSEClientTests.swift — helper / 응답 검증 케이스 제거 후 핵심 테스트만 유지
